Refactor move types to types folder
