+++ src/app/haku/[oid]/hakukohde/[hakukohde]/hakijaryhmat/hakijaryhma-content.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { Box } from '@mui/material';
-import { HakukohteenHakijaryhma } from '@/app/lib/laskenta-types';
+import { HakukohteenHakijaryhma } from '@/app/lib/types/laskenta-types';
 import React from 'react';
 import { HakijaryhmaAccordion } from './hakijaryhma-accordion';
 import { HakijaryhmaAccordionTitle } from './hakijaryhma-accordion-title';
