+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintalaskennan-tulos/valintatapajono-accordion-title.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { Box } from '@mui/material';
 import { useTranslations } from '@/app/hooks/useTranslations';
-import { LaskettuValinnanVaihe } from '@/app/lib/laskenta-types';
+import { LaskettuValinnanVaihe } from '@/app/lib/types/laskenta-types';
 import { Typography } from '@opetushallitus/oph-design-system';
 import React from 'react';
 import { toFormattedDateTimeString } from '@/app/lib/localization/translation-utils';
