+++ src/app/hooks/useLasketutValinnanVaiheet.ts
@@ -3,8 +3,8 @@ import { useSuspenseQueries } from '@tanstack/react-query';
 import { getLasketutValinnanVaiheet } from '../lib/valintalaskenta-service';
 import { getHakemukset } from '../lib/ataru';
 import { TranslatedName } from '../lib/localization/localization-types';
-import { JonoSija, LaskettuValintatapajono } from '../lib/laskenta-types';
-import { Hakemus } from '../lib/ataru-types';
+import { JonoSija, LaskettuValintatapajono } from '../lib/types/laskenta-types';
+import { Hakemus } from '../lib/types/ataru-types';
 
 export type JonoSijaWithHakijaInfo = Omit<
   JonoSija,
