+++ src/app/lib/types/kouta-types.ts
@@ -1,4 +1,4 @@
-import { TranslatedName } from './localization/localization-types';
+import { TranslatedName } from '../localization/localization-types';
 
 const STARTING_YEAR = 2019; // check earliest kouta haku
