+++ src/app/lib/types/sijoittelu-types.ts
